+++ Reflection/src/utopia/reflection/color/ColorRole.scala
@@ -5,10 +5,9 @@ package utopia.reflection.color
   * @author Mikko Hilpinen
   * @since 18.8.2020, v1.2
   */
-// TODO: Color role ought not to be sealed
-sealed trait ColorRole
+trait ColorRole
 
-sealed trait AdditionalColorRole extends ColorRole
+trait AdditionalColorRole extends ColorRole
 {
 	/**
 	  * @return A role that can be used instead of this color when another shade is required or this